requirements.txt
discord.py
guilded.py
discord-webhook
python-dotenv
webhooks